refactor: extract FutureWithAccumulation into its own file
